Refactor: Move `ChatAppSample` to a chatApp package
